networkDown case

+++ controller.runmean.cc
@@ -159,7 +159,7 @@ int Controller::chompWindow(int cint, double cwindDL) {
     //fprintf(fsend, "%lu: unseen last timestamp %lu = %lu\n", tStamp, lastAck, tStamp - lastAck );
     cint = 1;//cint/2;
   }
-  else if ((lastAck > 0)  && (cint == lastCW)) {
+  else if ((lastAck > 0)  && (cint == (int)lastCW)) {
     cint += 1; // don't ever stay in a state without exploring up
   }
   // make sure %change in cint isn't too spiky : causes delays
